bug fixed

+++ src/app/database/migrations/20251114070304_create_ig_latest_demographic_insights_table.js
@@ -18,7 +18,7 @@ module.exports = {
                 
                 percent_male_binary     FLOAT,
                 percent_female_binary   FLOAT,
-                binary_total            FLOAT,
+                binary_total            BIGINT,
                 
                 created_at TIMESTAMPTZ DEFAULT NOW(),
                 updated_at TIMESTAMPTZ DEFAULT NOW()
